move main layout to _app.js
